Add files via upload
As we all know, to maintain social distance is a need of an hour for fighting against COVID 19. So here is my project to track social distancing . You can also call it a social distancing alarm. Ultrasonic sensors send waves. These waves are absolutely invisible and come back after hitting an object. The Trig pin activates (D11) Or (D12). In which LED and Buzzer is connected. We have kept a distance of 100 cm